use isIntersecting instead of intersectionRatio for in-view detection
the previous check (entry.intersectionRatio > 0) can trigger false negatives,
as the value can be exactly 0 even as the element is coming into view;
it's not the property you are meant to use for this use case.

NOTE that the (entry.intersectionRatio > 0) check can be used safely,
as long as a `threshold: 0.01` is specified. I think this usage makes more sense,
and it leaves us to save the threshold to be used as a prop in the future,
if end user wants to have the trigger when
_element_ is 20% in view,
rather than using the inViewMargin (rootMargin), which is
element agnostic.